delete models; move lit_models -> models
