+++ tox.ini
@@ -18,9 +18,9 @@ deps =
     coverage
     hypothesis
 commands =
-    coverage run -m pytest -k "not test_litmodels.py"
+    coverage run -m pytest -k "not models"
     coverage report
-    pytest -k test_litmodels.py
+    pytest -k models
 
 [testenv:lint]
 deps =
